+++ src/Neo.Compiler.CSharp/MethodConvert/Helpers/SlotHelpers.cs
@@ -12,12 +12,10 @@ extern alias scfx;
 
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using scfx::Neo.SmartContract.Framework;
 using OpCode = Neo.VM.OpCode;
 
 namespace Neo.Compiler;
