+++ src/Neo.Compiler.CSharp/MethodConvert/Helpers/StackHelpers.cs
@@ -15,11 +15,55 @@ using Neo.VM;
 using System;
 using System.Buffers.Binary;
 using System.Numerics;
+using scfx::Neo.SmartContract.Framework;
+using OpCode = Neo.VM.OpCode;
 
 namespace Neo.Compiler;
 
 internal partial class MethodConvert
 {
+
+    #region Instructions
+    private Instruction AddInstruction(Instruction instruction)
+    {
+        _instructions.Add(instruction);
+        return instruction;
+    }
+
+    private Instruction AddInstruction(OpCode opcode)
+    {
+        return AddInstruction(new Instruction
+        {
+            OpCode = opcode
+        });
+    }
+
+    private SequencePointInserter InsertSequencePoint(SyntaxNodeOrToken? syntax)
+    {
+        return new SequencePointInserter(_instructions, syntax);
+    }
+
+    private SequencePointInserter InsertSequencePoint(SyntaxReference? syntax)
+    {
+        return new SequencePointInserter(_instructions, syntax);
+    }
+
+    private SequencePointInserter InsertSequencePoint(Location? location)
+    {
+        return new SequencePointInserter(_instructions, location);
+    }
+
+    #endregion
+
+    private Instruction Jump(OpCode opcode, JumpTarget target)
+    {
+        return AddInstruction(new Instruction
+        {
+            OpCode = opcode,
+            Target = target
+        });
+    }
+
     private void Push(bool value)
     {
         AddInstruction(value ? OpCode.PUSHT : OpCode.PUSHF);
@@ -208,5 +252,68 @@ internal partial class MethodConvert
         _breakTargets.Pop();
         if (_tryStack.TryPeek(out ExceptionHandling? result))
             result.BreakTargetCount--;
+    }
+
+
+    /// <summary>
+    /// Convert a throw expression or throw statement to OpCodes.
+    /// </summary>
+    /// <param name="model">The semantic model providing context and information about the Throw.</param>
+    /// <param name="exception">The content of exception</param>
+    /// <exception cref="CompilationException">Only a single parameter is supported for exceptions.</exception>
+    /// <example>
+    /// throw statement:
+    /// <code>
+    /// if (shapeAmount <= 0)
+    /// {
+    ///     throw new Exception("Amount of shapes must be positive.");
+    /// }
+    ///</code>
+    /// throw expression:
+    /// <code>
+    /// string a = null;
+    /// var b = a ?? throw new Exception();
+    /// </code>
+    /// <code>
+    /// var first = args.Length >= 1 ? args[0] : throw new Exception();
+    /// </code>
+    /// </example>
+    /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-throw-expression">The throw expression</seealso>
+    /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-try-catch-statement">Exception-handling statements - throw</seealso>
+    private void Throw(SemanticModel model, ExpressionSyntax? exception)
+    {
+        if (exception is not null)
+        {
+            var type = model.GetTypeInfo(exception).Type!;
+            if (type.IsSubclassOf(nameof(UncatchableException), includeThisClass: true))
+            {
+                AddInstruction(OpCode.ABORT);
+                return;
+            }
+        }
+        switch (exception)
+        {
+            case ObjectCreationExpressionSyntax expression:
+                switch (expression.ArgumentList?.Arguments.Count)
+                {
+                    case null:
+                    case 0:
+                        Push("exception");
+                        break;
+                    case 1:
+                        ConvertExpression(model, expression.ArgumentList.Arguments[0].Expression);
+                        break;
+                    default:
+                        throw new CompilationException(expression, DiagnosticId.MultiplyThrows, "Only a single parameter is supported for exceptions.");
+                }
+                break;
+            case null:
+                AccessSlot(OpCode.LDLOC, _exceptionStack.Peek());
+                break;
+            default:
+                ConvertExpression(model, exception);
+                break;
+        }
+        AddInstruction(OpCode.THROW);
     }
 }
